Logger: stars are working

+++ logger/logger_particle.h
@@ -74,7 +74,7 @@ struct logger_particle {
   float mass;
 
   /* unique id. */
-  size_t id;
+  long long id;
 
   /* time of the record. */
   double time;
